+++ src/index.ts
@@ -1,0 +1,7 @@
+"use strict";
+
+import {NodeSDK} from "./lib/NodeSDK";
+
+NodeSDK.NodeSDK = NodeSDK;
+module.exports = NodeSDK;
+export { NodeSDK as NodeSDK};