+++ src/lib/services/CallLogService.ts
@@ -1,5 +1,4 @@
 "use strict";
-import EventEmitter = NodeJS.EventEmitter;
 
 export {};
 
@@ -9,6 +8,7 @@ import {CallLogEventHandler} from '../connection/XMPPServiceHandler/calllogEvent
 //import {setFlagsFromString} from "v8";
 import {XMPPService} from "../connection/XMPPService";
 import {RESTService} from "../connection/RESTService";
+import {EventEmitter} from "events";
 import {isStarted} from "../common/Utils";
 import {Logger} from "../common/Logger";
 import {ContactsService} from "./ContactsService";
@@ -296,7 +296,7 @@ function CallLogsBean() : ICallLogsBean {
 
         // as duration is "h[H] mm[m] ss[s]" in rb, switch it back to ms ...
         for (let i = 0; i < callLogs.length; i++) {
-            let durationMs = 0;
+            let durationMs : any = 0;
             let hmmss = callLogs[i].duration;
             if (hmmss && (typeof hmmss === "string") && hmmss.match(/^(?:(?:([01]?\d|2[0-3])h )?([0-5]?\d)m )?([0-5]?\ds)$/)) {
                 // Remove h, m and s