+++ CHANGELOG.md
@@ -6,6 +6,26 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 Warning: Before deploying in production a bot that can generate heavy traffic, please contact ALE.
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+## [1.70.0-dotnet.0] - 2020-04-XX
+-   Add defaultDEBUG target in gruntfile to generate the compiled files with debug log not putted in comment.
+-   Fix generateRss file because Array.values() method did not work anymore.
+-   Update getContactByJid with a forceServerSearch parameter to force the search of the _contacts informations on the server.
+-   Add method `channelsService::likeItem` to like treatment/event on a Channel's Item with an appreciation
+-   Add method `channelsService::getDetailedAppreciations` to know in details apprecations given on a channel item (by userId the apprecation given)
+-   Add like events on a Channel's Item with an appreciation.
+-   Add in `postChangeLogInChannel` code to like the item when posting the changelog in Rainbow CPaaS info channel.
+-   Fix `postChangeLogInChannel` to wait until the publish and like item are done.
+-   Fix treatment of XMPP iq query event `set "remove"` from roster which was sending an "unavailable service" to server.
+-   Add method `contactsService::removeFromNetwork` to remove a contact from the list of contacts and unsubscribe to the contact's presence
+-   Add event `rainbow_contactremovedfromnetwork` raised when a contact is removed from connected user's network.
+-   Add in methods `ContactsService::getContactById`, `ContactsService::getContactByJid`, `ContactsService::getContactByLoginEmail`, `ContactsService::getContact` the presence if the requested user is the connected one.
+-   Add in Message received from server a field "geoloc: { datum: 'WGS84', latitude: '4x.567938', longitude: '-4.xxxxxxx' }" of the localisation sent in messages by a mobile.
+-   Move `index.js` to `src/index.ts` => become a typescript source file.
+-   Update libs to lastest version : `ncu -u` , so update sitemap library => fix sitemap_generation
+
+## [1.69.0] - 2020-03-30
+-   Add nextDebugBuildVersion program to generate to stdout a new debug version of  rainbow-node-sdk
+
 ## [1.69.0-dotnet.2] - 2020-03-23
 -   Same as 1.69.0-dotnet.0 (Failed of production)
 
@@ -17,9 +37,9 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Add document Connecting_to_Rainbow_S2S_Mode to describe the connection to Rainbow with a S2S event pipe.
 -   Official ChangeLog RSS Flow URL : https://hub.openrainbow.com/doc/sdk/node/api/ChangeLogRSS.xml
 -   Fix empty message list in conversation when conversationEventHandler::onConversationManagementMessageReceived event.
--   S2SService to service layer and folder
+-   Move S2SService to service layer and folder
 -   update S2S documentation. 
--   Add le parameter in joinroom in S2S
+-   Add the parameter role (Enum: "member" "moderator" of your role in this room) in joinroom in S2S
 -   Fix S2SServiceEventHandler::ParseRoomInviteCallback
 -   Fix error when delete conversation.